% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\encoding{utf-8}
\name{azteca}
\alias{azteca}
\title{An experiment on ant recruitment}
\format{A data frame with 21 rows and 3 variables:
\describe{
\item{plant}{plant id, integer}
\item{extract.new}{number of recruited ants in the leaf that received drops of smashed new leaves extract}
\item{extract.old}{number of recruited ants in the leaf that received drops of smashed old leaves extract}
}}
\source{
Kondrat, H. 2012. Estímulos químicos de folhas novas promovem recrutamento eficiente de formigas associadas à embaúba
\emph{Cecropia glaziovi} (Urticaceae). Curso de campo "Ecologia da Mata Atlântica"
(G. Machado; P.I. Prado & A.M.Z. Martini, eds.). Universidade de São Paulo, São Paulo.
\url{http://ecologia.ib.usp.br/curso/2012/PDF/PI-Hebert.pdf}
}
\usage{
azteca
}
\description{
Number of \emph{Azteca} ants recruited by leaf extracts of
their host plant , \emph{Cecropia} trees.
}
\section{Details}{

The ant colonies live in the hollow
trunk of \emph{Cecropia} and can detect and expel leaf-chewing insects.
To test if this response is more intense in young leaves, drops of extract of smashed young and old leaves
were poured in two neighbor leaves of the same plant. After 7 minutes the number of recruited ants in each leaf was recorded.
}
\keyword{datasets}

