% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{dplot}
\alias{dplot}
\title{Statistic distribution plot}
\usage{
dplot(dist, svalue, pside = c("Two sided", "Greater", "Lesser"),
  extreme = TRUE, vline = TRUE, rejection = TRUE, ...)
}
\arguments{
\item{dist}{the statistic distribution, as generated by
\code{\link{Rsampling}} (numeric vector)}

\item{svalue}{the result of applying the statistic over the original data}

\item{pside}{the alternative hypothesis for the hypothesis testing}

\item{extreme}{logical. should extreme points be highlighted in the plot?}

\item{vline}{logical. should the svalue be displayed as a vertical line?}

\item{rejection}{logical. should the critical region be highlighted?}

\item{...}{further arguments to be passed to \code{hist} function.}
}
\description{
Plots the distribution of the statistic of interest. Has switches to
plot the extreme values and
null hypothesis rejection region (also known as critical region).
}
\seealso{
See the package vignettes for more information about how to interpret this graph
}

