% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\encoding{utf-8}
\name{embauba}
\alias{embauba}
\title{Vine infestation on Cecropia trees}
\format{A data frame with 152 rows (plants) and 2 variables:
\describe{
\item{morphotype}{the tree morphotype, factor with two levels}
\item{with.vines}{does the tree harbor vines? Logical. }
}}
\source{
Mello, T.J. 2012. Infestação por lianas e comportamento de poda por formigas em \emph{Cecropia}
(Urticaceae). Curso de campo "Ecologia da Mata Atlântica"
(G. Machado; P.I. Prado & A.M.Z. Martini, eds.). Universidade de São Paulo, São Paulo.
\url{http://ecologia.ib.usp.br/curso/2012/PDF/PI-Thayna.pdf}
}
\usage{
embauba
}
\description{
Presence/absence data of vines on Cecropia trees of two morphotypes.
}
\section{Details}{

Two morphotypes of Cecropia trees differ in the occupancy by ant colonies.
Ants attack and drive out other insects that get to the trees.
To test if this protection also affect infestation by vines, trees of similar size of both morphs
were sampled and inspected for the presence of vines.
}
\keyword{datasets}

