% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\encoding{utf-8}
\name{peucetia}
\alias{peucetia}
\title{Preference of hunting spiders by hairy leaves}
\format{A data frame with 27 rows (trials) and 6 variables:
\describe{
\item{t1,t2,t3,t4,t5,t6}{Is the spider on the part covered by hairy leaves? Logical,
for each of 6 successive inspections (time 1, 2, ...}
}}
\source{
Werneck, R.T. 2010. Lar, viscoso lar. Experimento de seleção de habitat e forrageio de aranhas
em plantas com tricomas glandulares. Curso de campo "Ecologia da Mata Atlântica"
(G. Machado; P.I. Prado & A.A. Oliveira, eds.). Universidade de São Paulo, São Paulo.
\url{http://ecologia.ib.usp.br/curso/2010/pages/pdf/PI/relatorios/rachel.pdf}
}
\usage{
peucetia
}
\description{
Occupancy of \emph{Peucetia} spiders on parts of an experimental arena covered by
leaves with or without trichomes.
}
\section{Details}{

Spiders of the genus \emph{Peucetia} do not make webs and hunt actively on the vegetation.
The data is from an experiment to test if spiders prefer to stay in hairy leaves, that can stick their prey.
The spiders were kept in Petri dishes that had half of lower plate covered with hairy leaves.
The other half was covered by leaves without trichomes.
The placement of each spider was recorded 6 times at each 30 min.
}
\keyword{datasets}

