% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\encoding{utf-8}
\name{pielou}
\alias{pielou}
\title{Aphids recorded on goldenrods}
\format{A dataframe with 10 rows (aphid species of the genus Dactynotus) and 12 columns (plant species of the genus Solidago).
Each entry is the number of records of a given aphid species on a plant species.}
\source{
Pielou, E.C. 1972. Niche width and niche overlap: a method for measuring them. Ecology, 53: 687--692.
}
\usage{
pielou
}
\description{
Occurrences of aphids of the genus Dactynotus on plants of the genus Solidago in Canada
}
\section{Details}{

Data from a field survey by E.C. Pielou in Ontario to exemplify a method to calculate niche overlap and niche width.
The niche overlap gauges the overall similarity of the plant ranges used by the aphids. The niche width expresses how diverse
is the average diversity of plants used by the aphids.
}
\keyword{datasets}

