% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\encoding{utf-8}
\name{rhyzophora}
\alias{rhyzophora}
\title{Allometry in mangrove trees}
\format{A data frame with 24 rows (trees) and 4 variables:
\describe{
\item{soil.instability}{soil type according to instability; factor with two levels (high / medium)}
\item{canopy.trunk}{ratio between canopy and trunk area, both in m2, numeric}
\item{root}{area covered by aerial roots, numeric (m2)}
\item{n.roots}{number of aerial roots, integer}
}}
\source{
Prado, A. \emph{et al}. 2013. Variações na morfologia de sustentação em \emph{Rhizophora mangle} (Rizophoraceae) em diferentes
condições de inundação do solo. Curso de campo "Ecologia da Mata Atlântica"
(G. Machado, P.I. Prado & A.M.Z. Martini eds.). Universidade de São Paulo, São Paulo.
\url{http://ecologia.ib.usp.br/curso/2013/pdf/PO4-2.pdf}
}
\usage{
rhyzophora
}
\description{
Canopy to height ratio and variables of root area in mangrove trees sampled in two soil types.
}
\section{Details}{

Data from a field practical exercise to test if mangrove trees in more unstable soil
allocates more biomass in supporting roots.
}
\keyword{datasets}

