% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{samaras}
\alias{samaras}
\title{Distance traveled by winged fruits}
\format{A data frame with 30 rows (fruits) and 2 variables:
\describe{
\item{load}{wing load, the mass/area ratio (g/cm2)}
\item{with.vines}{distance traveled in a improvised wind tunnel (cm)}
}}
\source{
Cunha, T.J. 2011. Alometria de frutos alados e influência da frugivoria na capacidade de dispersão
de frutos de \emph{Stigmaphyllon} sp. (Malpighiaceae). Curso de campo "Ecologia da Mata Atlântica"
(G. Machado; P.I. Prado & A.M.Z. Martini, eds.). Universidade de São Paulo, São Paulo.
}
\usage{
samaras
}
\description{
Distance traveled in a wind tunnel by samaras with different area to mass ratios.
}
\keyword{datasets}

