% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{splot}
\alias{splot}
\title{Spaghetti plot}
\usage{
splot(p1, p2, highlight = TRUE, col.dif = c("black", "grey"), ...)
}
\arguments{
\item{p1, p2}{vectors of paired values (numerical vectors)}

\item{highlight}{should positive and negative differences within pairs highlighted with different colors? Logical}

\item{col.dif}{color vector if \code{highlight = TRUE}}

\item{...}{further arguments to be passed to plotting function (\code{stripchart}).}
}
\description{
Quick plot of paired differences, for exploratory purposes.
}

