% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{splot}
\alias{splot}
\title{Spaghetti plot}
\usage{
splot(p1, p2, highlight = TRUE, col.dif = c("black", "grey"),
  groups.names = c(1, 2), ...)
}
\arguments{
\item{highlight}{should positive and negative differences within pairs higlited with different colors? Logical}

\item{col.dif}{color vector if \code{highlight = TRUE}}

\item{groups.names}{labels for the groups names; numeric or character vector of length two.}

\item{...}{further arguments to be passed to \code{plot} function.
(e.g., \code{cols}, \code{replace}, \code{stratum}).}

\item{{p1,}{p2} vectors of paired values; numerical vectors}
}
\description{
Quick plot of paired differences, for exploratory purposes.
}

